Add td-lean.hh, fix td::future move assign
